requirements.txt
asgiref==3.8.1
CacheControl==0.14.0
certifi==2024.7.4
cffi==1.17.0
charset-normalizer==3.3.2
cryptography==43.0.0
distlib==0.3.8
Django==5.1
django-cors-headers==4.4.0
django-dotenv==1.4.2
django-filter==24.3
djangorestframework==3.15.2
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
idna==3.7
importlib_metadata==8.2.0
inflection==0.5.1
installer==0.7.0
jaraco.classes==3.4.0
more-itertools==10.4.0
msgpack==1.0.8
mysqlclient==2.2.4
Naked==0.1.32
numpy==2.0.1
openpyxl==3.1.5
packaging==24.1
pandas==2.2.2
pexpect==4.9.0
ptyprocess==0.7.0
pycparser==2.22
PyJWT==2.9.0
pyproject_hooks==1.1.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.1
PyYAML==6.0.2
requests==2.32.3
simplejwt==2.0.1
six==1.16.0
sqlparse==0.5.1
status==0.2.5
typing==3.7.4.3
tzdata==2024.1
uritemplate==4.1.1
urllib3==2.2.2
XlsxWriter==3.2.0
zipp==3.20.0
